refactor(postman-to-openapi): export isPostmanCollection helper
