+++ packages/api-client/src/v2/features/command-palette/helpers/load-document-from-source.ts
@@ -1,10 +1,10 @@
 import { isObject } from '@scalar/helpers/object/is-object'
 import type { LoaderPlugin } from '@scalar/json-magic/bundle'
 import { parseJson, parseYaml } from '@scalar/json-magic/bundle/plugins/browser'
+import { isPostmanCollection } from '@scalar/postman-to-openapi'
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
 
 import { getOpenApiFromPostman } from '@/v2/features/command-palette/helpers/get-openapi-from-postman'
-import { isPostmanCollection } from '@/v2/features/command-palette/helpers/is-postman-collection'
 
 export type ImportEventData = {
   source: string
