+++ packages/postman-to-openapi/src/index.ts
@@ -1,2 +1,3 @@
 export { type ConvertOptions, type PostmanRequestIndexPath, convert } from './convert'
+export { isPostmanCollection } from './is-postman-collection'
 export { extractPathFromUrl, normalizePath } from './helpers/urls'
